refactor(frontend): partir admin-dashboard en 3 sub-componentes (fase 5b-1)
Padre admin-dashboard.component.ts pasa de 558 a 174 lineas (TS) y de 474 a 66 (HTML).
Pasa a ViewEncapsulation.None para que los estilos del padre alcancen a los hijos.

Sub-componentes creados (todos standalone, autosuficientes):
- app-admin-panel     26 TS + 112 HTML  stats + previews + quick actions
- app-admin-usuarios  258 TS + 183 HTML tabla + tabs + CRUD alumnos/doctores + modal confirmar
- app-admin-calendario 123 TS + 116 HTML calendario + form dia especial + lista del mes

Helpers compartidos:
- admin-utils.ts  91  UsuarioRow, CalendarCellAdmin, formatFechaDia, toAlumnoRow,
                     toDoctorRow, buildCalGridAdmin, filterDiasProximos, todayYmd

Padre conserva: topbar + sidebar + ruteo entre secciones + cache de panel data.
Sub-componentes emiten (changed) para que padre actualice las previews del panel.

ng build OK. Sin cambios funcionales.
