+++ frontend/src/app/components/admin/admin-dashboard/calendario/admin-calendario.component.html
@@ -1,0 +1,116 @@
+<section class="section">
+  <div class="cal-cols">
+
+    <div class="cal-col-left">
+      <div class="card">
+        <div class="cal-nav">
+          <button class="cal-nav-btn" (click)="changeCalMonth(-1)">&#8249;</button>
+          <span class="cal-month-label">{{ calLabel | titlecase }}</span>
+          <button class="cal-nav-btn" (click)="changeCalMonth(1)">&#8250;</button>
+        </div>
+
+        @if (isLoadingCal) {
+          <div class="loading">Cargando...</div>
+        } @else {
+          <div class="cal-weekdays-6">
+            @for (d of weekDays; track d) { <span>{{ d }}</span> }
+          </div>
+          <div class="cal-weeks-6">
+            @for (week of calWeeks; track $index) {
+              @for (day of week; track day.date) {
+                <div
+                  class="cal-day"
+                  [class.other-month]="!day.isCurrentMonth"
+                  [class.past]="day.isPast && !day.diaEspecial"
+                  [class.selected]="diaForm.fecha === day.date"
+                  [class.fes]="day.diaEspecial?.tipo === 'holiday'"
+                  [class.vac]="day.diaEspecial?.tipo === 'vacation'"
+                  [class.red]="day.diaEspecial?.tipo === 'reduced'"
+                  (click)="day.isCurrentMonth && selectCalDay(day.date, day.isPast, !!day.diaEspecial)"
+                  [title]="day.diaEspecial?.etiqueta ?? (day.isPast && !day.diaEspecial ? 'Día pasado — no editable' : day.isCurrentMonth ? 'Clic para marcar' : '')">
+                  <span class="day-num">{{ day.label }}</span>
+                  @if (day.diaEspecial) {
+                    <span class="day-dot" [class]="day.diaEspecial.tipo"></span>
+                  }
+                </div>
+              }
+            }
+          </div>
+        }
+
+        <div class="cal-legend">
+          <span class="leg-item fes">Festivo</span>
+          <span class="leg-item vac">Vacaciones</span>
+          <span class="leg-item red">Reducido</span>
+        </div>
+      </div>
+    </div>
+
+    <div class="cal-col-right">
+      <div class="card">
+        <div class="card-head">
+          <span>{{ diaForm.fecha ? 'Editar día · ' + formatFechaDia(diaForm.fecha) : 'Agregar día especial' }}</span>
+          @if (diaForm.fecha) {
+            <button class="preview-link" (click)="resetDiaForm()">Limpiar</button>
+          }
+        </div>
+        @if (diaMsg) {
+          <div class="form-msg" [class.success]="diaMsg.includes('correct')">{{ diaMsg }}</div>
+        }
+        <div class="form-group">
+          <label>Fecha *</label>
+          <input type="date" [(ngModel)]="diaForm.fecha" class="form-input">
+        </div>
+        <div class="form-group" style="margin-top: 12px">
+          <label>Tipo *</label>
+          <select [(ngModel)]="diaForm.tipo" class="form-select">
+            <option value="holiday">Festivo / Inhábil</option>
+            <option value="vacation">Vacaciones</option>
+            <option value="reduced">Día reducido (horario parcial)</option>
+          </select>
+        </div>
+        <div class="form-group" style="margin-top: 12px">
+          <label>Etiqueta</label>
+          <input type="text" [(ngModel)]="diaForm.etiqueta" placeholder="Ej: Semana Santa, Puente..." class="form-input">
+        </div>
+        @if (diaForm.tipo === 'reduced') {
+          <div class="form-group" style="margin-top: 12px">
+            <label>Hora de cierre *</label>
+            <input type="time" [(ngModel)]="diaForm.hora_cierre" min="08:00" max="16:45" class="form-input">
+            <small class="form-hint">A partir de esta hora no se podrán agendar citas.</small>
+          </div>
+        }
+        <button class="btn-pri btn-full" style="margin-top: 16px" (click)="submitDia()" [disabled]="isSubmittingDia">
+          {{ isSubmittingDia ? 'Guardando...' : 'Guardar día' }}
+        </button>
+      </div>
+
+      @if (diasEspeciales.length > 0) {
+        <div class="card" style="margin-top: 16px">
+          <div class="card-head"><span>Días registrados este mes</span></div>
+          @for (dia of diasEspeciales; track dia.id) {
+            <div class="dia-reg-item">
+              <div class="dia-reg-fecha">{{ formatFechaDia(dia.fecha) }}</div>
+              <span class="dia-preview-tipo" [class]="dia.tipo">{{ tipoLabels[dia.tipo].label }}</span>
+              <span class="dia-reg-etiqueta">
+                {{ dia.etiqueta || '—' }}
+                @if (dia.tipo === 'reduced' && dia.hora_cierre) {
+                  <small class="dia-reg-hora">· cierre {{ dia.hora_cierre }}</small>
+                }
+              </span>
+              <div class="dia-reg-actions">
+                <button class="action-link" (click)="selectCalDay(dia.fecha)">Editar</button>
+                <button class="action-link warn" (click)="deleteDia(dia.id)">Eliminar</button>
+              </div>
+            </div>
+          }
+        </div>
+      }
+
+      @if (diasEspeciales.length === 0 && !isLoadingCal) {
+        <div class="empty-state" style="margin-top: 16px">Sin días especiales este mes.</div>
+      }
+    </div>
+
+  </div>
+</section>
